chore(swap): restructure form directory
Just moved form top-section to a dedicated directory and moved form related
stuff from swap directory to the form subdirectory.
