+++ src/renderer/screens/exchange/Swap2/Form/FormSelectors/index.js
@@ -1,0 +1,1 @@
+export { default } from "./FormInputs";
